4: Refactor how Java 14 is specified (#3)
After testing, and also getting JAVA_HOME set to the right JDK in
/etc/profile.d/jdk.sh (on Ubuntu 20.04) then I could be much more hands
off with regard to specifying the java version. I no longer need to have
the maven-complier-plugin specify which javac compiler to use!

Removed the import of the securbanking-bom

After discussion with the team it was felt that the best approach would
be to not include all versions in an imported parent bom. This leads to
circular dependencies when you update the bom with latest versions,
update the parent pom to use the latest bom, then need to update all the
child projects that use the parent... not good.

So instead we will use a bom file in the securebanking-commons library
that may be imported into clients such as the securebanking-rcs or
securebanking-openbanking-aspsp projects.

Issue: https://github.com/SecureBankingAcceleratorToolkit/SecureBankingAcceleratorToolkit/issues/4